+++ src/app/dashboard/applications/components/application-panel/application-panel.module.ts
@@ -8,7 +8,6 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { ApplicationPanelComponent } from './application-panel.component';
 
@@ -32,7 +31,6 @@ import { DayDifferencePipeModule } from '~pipes/day-difference/day-difference.pi
     MatInputModule,
     MatMenuModule,
     MatProgressSpinnerModule,
-    MatTooltipModule,
     ReactiveFormsModule,
     TagModule
   ],