+++ src/app/dashboard/dashboard.module.ts
@@ -2,19 +2,13 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatRippleModule } from '@angular/material/core';
-import { MatDialogModule } from '@angular/material/dialog';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatIconModule } from '@angular/material/icon';
-import { MatListModule } from '@angular/material/list';
-import { MatSelectModule } from '@angular/material/select';
 import { MatSidenavModule } from '@angular/material/sidenav';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { RouterModule } from '@angular/router';
 
 import { DashboardRoutingModule } from './dashboard-routing.module';
 import { DashboardComponent } from './dashboard.component';
 
-import { HelpModule } from '~components/help/help.module';
 import { JobBoardSelectorModule } from '~components/job-board-selector/job-board-selector.module';
 import { ToolbarModule } from '~components/toolbar/toolbar.module';
 
@@ -23,18 +17,12 @@ import { ToolbarModule } from '~components/toolbar/toolbar.module';
   imports: [
     CommonModule,
     DashboardRoutingModule,
-    HelpModule,
     JobBoardSelectorModule,
     MatButtonModule,
-    MatDialogModule,
     MatDividerModule,
     MatIconModule,
-    MatListModule,
-    MatSelectModule,
     MatSidenavModule,
-    MatTooltipModule,
     MatRippleModule,
-    RouterModule,
     ToolbarModule
   ]
 })