+++ src/app/dashboard/job-boards/components/job-board-selector/job-board-selector.module.ts
@@ -1,15 +1,11 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatDividerModule } from '@angular/material/divider';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatSelectModule } from '@angular/material/select';
 
 import { JobBoardSelectorComponent } from './job-board-selector.component';
 
@@ -19,18 +15,13 @@ import { JobBoardDialogModule } from '~components/job-board-dialog/job-board-dia
   declarations: [JobBoardSelectorComponent],
   imports: [
     CommonModule,
-    FormsModule,
     MatButtonModule,
     MatDialogModule,
     MatDividerModule,
-    MatFormFieldModule,
     MatIconModule,
-    MatInputModule,
     MatMenuModule,
     MatProgressSpinnerModule,
-    MatSelectModule,
-    JobBoardDialogModule,
-    ReactiveFormsModule
+    JobBoardDialogModule
   ],
   exports: [JobBoardSelectorComponent]
 })