+++ src/app/shared/components/help/help.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'at-help',
-  templateUrl: './help.component.html',
-  styleUrls: ['./help.component.scss']
-})
-export class HelpComponent {}